+++ docs/changelog.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+Changes in v0.5.0
+=================
+- Add ability to set ``allow_blank_source`` on serializer class. If this is set to ``True`` the serializer class will not throw error if source is not available.
+
+Changes in v0.4.1
+=================
+- Fix bug where nested serialization where caching data.
+
 Changes in v0.4.0
 =================
 - Add support for ``MethodField``